feat: create ProjectLayout component and refactor ProjectPage to use mock tasks
